refactor functions for plot-grammar, i.e. move code scattered across files into one new plot_grammar.cc file
